% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonenburg.R
\docType{data}
\name{bonenburg}
\alias{bonenburg}
\title{Bonenburg geochemical dataset}
\format{
A data frame with 81 rows and 11 variables:
\describe{
  \item{Section}{Name of the section}
  \item{Strat}{First order stratigraphic units}
  \item{Strat2}{Second order stratigraphic units}
  \item{SampleID}{Assigned IDs}
  \item{Height}{Height in units cm}
  \item{CaCO3}{Calcium carbonate content in units weight percent of rock}
  \item{TN}{Total nitrogen in units weight percent of rock}
  \item{del13Ctoc}{Carbon isotope composition of total organic carbon in units per mille relative to Vienna Peedee Belemnite}
  \item{TOCcfb}{Total organic carbon on carbonate free basis in units weight percent of rock}
  \item{Na_Al}{Sodium to aluminium ratio}
  \item{K_Al}{Potassium to aluminium ratio}
}
}
\source{
https://www.schweizerbart.de/papers/nos/detail/52/90527/A_comparative_study_of_total_organic_carbon_d13C_signatures_in_the_Triassic_Jurassic_transitional_beds_of_the_Central_European_Basin_and_western_Tethys_shelf_seas
}
\usage{
bonenburg
}
\description{
Bonenburg geochemical dataset
}
\keyword{datasets}
