% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonenburg_long.R
\docType{data}
\name{bonenburg_long}
\alias{bonenburg_long}
\title{Bonenburg geochemical dataset long format}
\format{
A data frame with 486 rows and 11 variables:
\describe{
\item{Section}{Name of the section}
\item{Strat}{First order stratigraphic units}
\item{Strat2}{Second order stratigraphic units}
\item{SampleID}{Assigned IDs}
\item{Height}{Height in units cm}
\item{measurement}{Type of measurement}
\item{value}{Value of measurment}
}
}
\source{
https://www.schweizerbart.de/papers/nos/detail/52/90527/A_comparative_study_of_total_organic_carbon_d13C_signatures_in_the_Triassic_Jurassic_transitional_beds_of_the_Central_European_Basin_and_western_Tethys_shelf_seas
}
\usage{
bonenburg_long
}
\description{
This is the same dataset but now the continuous variables (e.g. measurements)
are crossed with the variable Height.
}
\keyword{datasets}
