% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonenburg_tidy.R
\docType{data}
\name{bonenburg_tidy}
\alias{bonenburg_tidy}
\title{Bonenburg geochemical dataset}
\format{
A data frame with 81 rows and 12 variables:
\describe{
  \item{Section}{Name of the section}
  \item{Strat}{First order stratigraphic units}
  \item{Strat2}{Second order stratigraphic units}
  \item{SampleID}{Assigned IDs}
  \item{Height}{Height in units cm}
  \item{CaCO3}{Calcium carbonate content in units weight percent of rock}
  \item{TN}{Total nitrogen in units weight percent of rock}
  \item{del13Ctoc}{Carbon isotope composition of total organic carbon in units per mille relative to Vienna Peedee Belemnite}
  \item{TOCcfb}{Total organic carbon on carbonate free basis in units weight percent of rock}
  \item{Al2O3_pc}{XRF Aliminium oxide in units weight percent of rock}
  \item{Na2O_pc}{XRF Sodium oxide in units weight percent of rock}
  \item{K2O_pc}{XRF Potasium oxide in units weight percent of rock}
}
}
\source{
https://www.schweizerbart.de/papers/nos/detail/52/90527/A_comparative_study_of_total_organic_carbon_d13C_signatures_in_the_Triassic_Jurassic_transitional_beds_of_the_Central_European_Basin_and_western_Tethys_shelf_seas
}
\usage{
bonenburg_tidy
}
\description{
This is the same dataset but the XRF data have not been converted to
elemental ratios
}
\keyword{datasets}
