% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasaurus_dozen.R
\docType{data}
\name{datasaurus_dozen}
\alias{datasaurus_dozen}
\title{datasauRus dozen}
\format{
A data frame with 1846 rows and 3 variables:
\describe{
\item{dataset}{indicates which dataset the data are from}
\item{x}{x-values}
\item{y}{y-values}
}
}
\usage{
datasaurus_dozen
}
\description{
See \code{datasauRus::\link[datasauRus:datasaurus_dozen]{datasaurus_dozen}} for details.
}
\keyword{datasets}
