% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{dplyr}
\alias{dplyr}
\alias{mutate}
\alias{group_by}
\title{dplyr functions used in PAGES seminar}
\description{
\itemize{
\item \code{\link[dplyr:group_by]{group_by}()} Converts a
\code{\link[tibble:tibble]{tibble}} to a grouped \code{tibble} where operations
are performed group-wise.
\item \code{\link[dplyr:mutate]{mutate}()} Transform a variable of a
\code{\link[tibble:tibble]{tibble}}
}

To learn more about dplyr visit \url{https://r4ds.had.co.nz/index.html}.
}
\keyword{internal}
