% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot2}
\alias{ggplot2}
\alias{ggplot}
\alias{aes}
\alias{facet_wrap}
\alias{facet_grid}
\alias{geom_point}
\alias{geom_area}
\alias{geom_boxplot}
\alias{geom_smooth}
\alias{vars}
\title{Common ggplot functions used in PAGES seminar}
\description{
\itemize{
\item \code{\link[ggplot2:ggplot]{ggplot}()} initializes a ggplot object.
\item \code{\link[ggplot2:aes]{aes}()} Aesthetic mappings describe how
variables in the data are mapped to visual properties (aesthetics) of
geoms.
\item \code{\link[ggplot2:geom_point]{geom_point}()} The point geom is used to
create scatterplots.
\item \code{\link[ggplot2:geom_area]{geom_area}()} For each x value,
\code{geom_area()} displays a y interval defined by 0 and ymax (or 0 and xmax).
\item \code{\link[ggplot2:geom_smooth]{geom_smooth}()} For detecting patterns in
data.
\item \code{\link[ggplot2:geom_boxplot]{geom_boxplot}()} The boxplot compactly
displays the distribution of a continuous variable.
\item \code{\link[ggplot2:facet_grid]{facet_grid}()} forms a matrix of panels
defined by row and column faceting variables.
\item \code{\link[ggplot2:facet_wrap]{facet_wrap}()} wraps a 1d sequence of
panels into 2d.
\item \code{\link[ggplot2:vars]{vars}()} Just like \code{aes()}, \code{vars()} is a quoting
function that takes inputs to be evaluated in the context of a dataset.
}

To learn more about ggplot2 \url{https://r4ds.had.co.nz/index.html} and
\url{http://www.cookbook-r.com/Graphs/}
}
\keyword{internal}
