% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kuhjoch_long.R
\docType{data}
\name{kuhjoch_long}
\alias{kuhjoch_long}
\title{Kuhjoch palynological dataset long format}
\format{
A data frame with 180 rows and 4 variables:
\describe{
\item{Height}{Height in units cm}
\item{type}{Characteristic of palynomorph assemblage type}
\item{label}{Type of palynomorph assemblage}
\item{count}{Count}
}
}
\source{
https://www.schweizerbart.de/papers/nos/detail/52/90527/A_comparative_study_of_total_organic_carbon_d13C_signatures_in_the_Triassic_Jurassic_transitional_beds_of_the_Central_European_Basin_and_western_Tethys_shelf_seas
}
\usage{
kuhjoch_long
}
\description{
Kuhjoch palynological dataset long format
}
\keyword{datasets}
