% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAGES.R
\docType{package}
\name{PAGES}
\alias{PAGES}
\title{A package for the PAGES ECN webinar: First steps in exploring data with R.}
\description{
The PAGES package provides stratigraphic datasets and documentation to
explore tidyverse functions.
}
