% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot.R
\name{pivot_longer}
\alias{pivot_longer}
\title{pivot_longer used in PAGES seminar}
\description{
\code{\link[tidyr:pivot_longer]{pivot_longer}()} Converts a wide format
\code{\link[tibble:tibble]{tibble}} to a long format \code{tibble}.

To learn more about tidyr visit \url{https://r4ds.had.co.nz/index.html}.
}
\keyword{internal}
