% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_slides}
\alias{render_slides}
\title{Render slides of PAGES presentation}
\usage{
render_slides(output_dir = ".")
}
\arguments{
\item{output_dir}{where to locate the output HTML file.}
}
\description{
This function allows rendering of the rmarkdown ioslides used for the
presentation
}
